requirements.txt
torch>=2.0.0
pytorch-ignite>=0.5.0
numpy
tqdm
requests